models/SEED/requirements.txt
xformers>=0.0.20
omegaconf
hydra-core
pyrootutils
transformers==4.30.2
flask
gradio
timm
diffusers==0.20.2
accelerate
einops
sentencepiece
gradio==3.50.0
